Fix terminal dimension type consistency and add comprehensive validation tests
- Fix u16/u32 type mismatch between global and subcommand terminal dimensions
- Add comprehensive integration tests for all validation scenarios
- Tests verify exact error messages per spec
- All existing tests pass with the type fixes

+++ crates/deacon/src/cli.rs
@@ -278,10 +278,10 @@ pub enum Commands {
         skip_feature_auto_mapping: bool,
         /// Terminal columns (requires --terminal-rows)
         #[arg(long)]
-        terminal_columns: Option<u16>,
+        terminal_columns: Option<u32>,
         /// Terminal rows (requires --terminal-columns)
         #[arg(long)]
-        terminal_rows: Option<u16>,
+        terminal_rows: Option<u32>,
         /// User data folder (accepted but not used by this subcommand)
         #[arg(long)]
         user_data_folder: Option<PathBuf>,
